2023.02 V1.0 LTS

+++ TI3/Assets/Scripts/Route/UIRoute.cs
@@ -5,8 +5,8 @@ using UnityEngine.UI;
 
 public class UIRoute : MonoBehaviour
 {
-    public int firstColonyIndex;
-    public int secondColonyIndex;
+    public sbyte firstColonyIndex;
+    public sbyte secondColonyIndex;
     [HideInInspector] public Image image; // Button image
     private void Awake() { image = this.GetComponentInChildren<Image>(); }
     public void OnButtonSelectRoute()
